requirements.txt
pandas < 1.2.0

xlrd < 1.2.0

openpyxl < 3

streamlit < 1.3

plotly < 4.14

Pillow < 8.1.2